Map no longer needs to calculate a special visual center
Since the left half of the map is no longer hidden behind the
sidebar, the map no longer needs to use fixmystreet_midpoint()
etc to calculate the "visual" centre of the map. OpenLayers’
built-in centering functions will do the job.

We also remove midpoint_box_excluding_column because the
.report-a-problem-sidebar will hopefully also be going, so
it’s another thing that the centering functions can ignore.