Show actual report destination in confirm message
For two-tier authorities, it is possible to send a report
from a cobrand to an authority other than the cobrand's owner.

If this is the case, then highlight it on the confirmation message
along with a link to view the problem report on the main fixmystreet.com
site.

The Hart-specific template is no longer required.