Centre map on new issue location when creating a report
Now that it’s possible for the map element to change width to
accommodate the .with-notes variant (eg: in Zurich), the point
you selected could be repositioned outside the viewport.

This stops that happening by always centering the map on the
new point when creating a new report. It feels very natural,
so I’ve made it the default behaviour for all cobrands,
whether they use .with-notes or not.